+++ sys/src/9/mtx/mmu.c
@@ -235,7 +235,7 @@ putmmu(ulong va, ulong pa, Page *pg)
 }
 
 void
-checkmmu(ulong, ulong)
+checkmmu(uintptr, uintptr)
 {
 }
