+++ sys/src/9/pc/dat.h
@@ -38,7 +38,7 @@ struct Lock
 {
 	ulong	key;
 	ulong	sr;
-	ulong	pc;
+	uintptr	pc;
 	Proc	*p;
 	Mach	*m;
 	ushort	isilock;
